+++ src/hooks/useMatch.ts
@@ -1,12 +1,12 @@
 import { useEffect, useState } from "react";
-import { getMatchday } from "../api";
+import { getMatch } from "../api";
 import { LiveBlogEntryItem } from "../types/firebase";
 
-export const useMatchday = (url: string) => {
+export const useMatch = (url: string) => {
   const [entries, setEntries] = useState<LiveBlogEntryItem[]>();
 
   useEffect(() => {
-    getMatchday(url).then((data) => {
+    getMatch(url).then((data) => {
       setEntries(data ? Object.values(data) : []);
     });
   }, [url]);